Move to/from string methods to the json schema package. Delete utils.go from template library
